Commit voor TramController en Layout

+++ Eyect4RailsWebApp/Repositories/MSSQLRepository/MSSQLTramRepository.cs
@@ -100,7 +100,7 @@ namespace Eyect4RailsWebApp.Repositories.MSSQLRepository
                         {
                             command.Parameters.AddWithValue("@ID", entity.Id);
                             command.Parameters.AddWithValue("@RemiseID", entity.RemiseId);
-                            command.Parameters.AddWithValue("@Tramtype", entity.Tramtype);
+                            command.Parameters.AddWithValue("@Tramtype", (int)entity.Tramtype);
                             command.Parameters.AddWithValue("@Tramnumber", entity.TramNumber);
                             command.Parameters.AddWithValue("@Length", entity.Length);
                             command.Parameters.AddWithValue("@Filthy", entity.Filthy);
